Moved from delayed to lazy, to capture navigation analytics
