+++ scripts/scripts.js
@@ -187,6 +187,7 @@ async function loadLazy(doc) {
 
   loadCSS(`${window.hlx.codeBasePath}/styles/lazy-styles.css`);
   loadFonts();
+  import('./lazy.js');
 }
 
 /**
